% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sampler.R
\name{log_partial_likelihood}
\alias{log_partial_likelihood}
\title{Compute the log partial-likelihood}
\usage{
log_partial_likelihood(Y, delta, X, b)
}
\arguments{
\item{Y}{observed failure / censoring times}

\item{delta}{censoring vector (0: censored, 1: uncensored)}

\item{X}{matrix of observations}

\item{b}{vector of coefficients}
}
\description{
Compute the log partial-likelihood
}
