% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sampler.R
\name{run_sampler}
\alias{run_sampler}
\title{Run MCMC sampler}
\usage{
run_sampler(
  Y,
  delta,
  X,
  lambda,
  kernel_sd = 0.2,
  mcmc_samples = 5000,
  verbose = T
)
}
\arguments{
\item{Y}{observed failure / censoring times}

\item{delta}{censoring vector (0: censored, 1: uncensored)}

\item{X}{matrix of observations}

\item{lambda}{penalisation parameter}

\item{kernel_sd}{standard deviation of sampling kernel}

\item{mcmc_samples}{number of samples}

\item{verbose}{print a progress bar}
}
\description{
Run MCMC sampler
}
\examples{
set.seed(1)
n <- 100; omega <- 1; censoring_lvl <- 0.4
b <- c(2, 2, rep(0, 148)); p <- length(b)
X <- matrix(rnorm(n * p), nrow=n)
y <- runif(nrow(X))
Y <- log(1 - y) / - (exp(X \%*\% b) * omega)
delta  <- runif(n) > censoring_lvl      # 0: censored, 1: uncensored
Y[!delta] <- Y[!delta] * runif(sum(!delta))

run_sampler(Y, delta, X, 2)

}
